// app/api/generated/models/CustomerProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Amra Merch Market API
 * Contract-first API for Amra Shop.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CustomerAddress } from './CustomerAddress';
import {
    CustomerAddressFromJSON,
    CustomerAddressFromJSONTyped,
    CustomerAddressToJSON,
    CustomerAddressToJSONTyped,
} from './CustomerAddress';

/**
 *
 * @export
 * @interface CustomerProfile
 */
export interface CustomerProfile {
    /**
     *
     * @type {string}
     * @memberof CustomerProfile
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof CustomerProfile
     */
    phone: string;
    /**
     *
     * @type {string}
     * @memberof CustomerProfile
     */
    displayName?: string | null;
    /**
     *
     * @type {string}
     * @memberof CustomerProfile
     */
    email?: string | null;
    /**
     *
     * @type {boolean}
     * @memberof CustomerProfile
     */
    emailVerified: boolean;
    /**
     *
     * @type {Array<CustomerAddress>}
     * @memberof CustomerProfile
     */
    addresses: Array<CustomerAddress>;
}

/**
 * Check if a given object implements the CustomerProfile interface.
 */
export function instanceOfCustomerProfile(value: object): value is CustomerProfile {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('phone' in value) || value['phone'] === undefined) return false;
    if (!('emailVerified' in value) || value['emailVerified'] === undefined) return false;
    if (!('addresses' in value) || value['addresses'] === undefined) return false;
    return true;
}

export function CustomerProfileFromJSON(json: any): CustomerProfile {
    return CustomerProfileFromJSONTyped(json, false);
}

export function CustomerProfileFromJSONTyped(json: any, ignoreDiscriminator: boolean): CustomerProfile {
    if (json == null) {
        return json;
    }
    return {

        'id': json['id'],
        'phone': json['phone'],
        'displayName': json['displayName'] == null ? undefined : json['displayName'],
        'email': json['email'] == null ? undefined : json['email'],
        'emailVerified': json['emailVerified'],
        'addresses': ((json['addresses'] as Array<any>).map(CustomerAddressFromJSON)),
    };
}

export function CustomerProfileToJSON(json: any): CustomerProfile {
    return CustomerProfileToJSONTyped(json, false);
}

export function CustomerProfileToJSONTyped(value?: CustomerProfile | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'id': value['id'],
        'phone': value['phone'],
        'displayName': value['displayName'],
        'email': value['email'],
        'emailVerified': value['emailVerified'],
        'addresses': ((value['addresses'] as Array<any>).map(CustomerAddressToJSON)),
    };
}